Fix secondary index queries with paging
The assertion in SecondaryIndexSearcher.makeIndexFilter that if the left
side of a DataRange is a DecoratedKey, then so must the right side be
was causing some 2i queries to fail after the first page. It's possible
then for range.left to be the last seen key and range.right to be the
min token.

Conflicts:
	src/java/org/apache/cassandra/db/index/SecondaryIndexSearcher.java