Properly extract whether a legacy slice queries static columns
Pre-3.0 nodes were using the slices to query or not static columns. 3.0
handles this though the ColumnFilter (by explicitly asking for
static columns or not) and has no real way to do so through its slices.

The backward compatibility code was not handling this correctly: it was
generating the 3.0 slices from the pre-3.0 one somewhat blindly, and
then checking if said slices were including Clustering.STATIC. This
work most of the time because most query selecting static simply
started at Slice.Bound.BOTTOM which ended up selecting
Clustering.STATIC, but wasn't fully correct.

The patch makes sure we detect if static are selected when
deserializing the old slices.