Faster UUID->ByteBuffer
We convert UUID to their ByteBuffer representation in a few places and
do that by converting the UUID to a byte[] and then wrapping it. But
our convertion of UUID to byte[] writes the 2 long the UUID is composed
of byte by byte, while for a ByteBuffer we can easily just "put" both
long. So the patch introduces a specific method to do that. That's
arguably cleaner anyway.