Memoize hash value for DataResource
Every request does a hash map lookup on a DataResource (for validating
access rights) which imply calling DataResource.hashCode(). Currently
that uses Objects.hashCode(), which is a varargs methods and allocate
an array (which, according to MissionControl seems to not be stack
allocated).

The patch does 2 things:
 1) it computes the hash only once in the ctor
 2) it stores the DataResource for each table in the CFMetaData and uses
    that when checking access permissions for SelectStatement and
    ModificationStatement (the 2 statements where performance matters).
The 2nd point also saves some unecessary test that the table exists when
we actually know it does.