Fix regression from CASSANDRA-10136
When a read crossed a column index boundary (reading old sstable), we
have to mark the beginning of the index block and so we seek back
to said index block start, mark, and seek back to where we were to
continue reading. The problem was that the seekToPosition() method was
clearing the deserialization state which, in that case, is not what
we wanted.

The patch move the clearing out of the seek method and only call it
when that's actually what we want to do.