Move handshake message details to a separate class
The main purpose is just to pull the message details to a single place
so the protocol is more easily followed (and make things a bit cleaner
I feel as a result). This also allow to document it closer to the code
which is probably not a bad thing.

Side note: while this does create new objects, 1) the handshake isn't
that performance sensitive and 2) they are probably stack allocated
anyway.