Move coalescing/flush related behavior to single class
The main part of this patch is the introduction of the OutChannel class. It
directly replaces the OutboundMessagingConnection.ChannelWrapper class, but more
importantly it attempts to group everything related to flush and coalescing in
a single class to hopefully make it clearer what we do exactly.

Another thing that is moved out of MessageOutChannel and into this new class
is the counting of messages (proceed, dropped and pending). Having those updated
"far" from where they were use felt confusing to me, and in particular for
pending ones, we relied on each increment being met by a following decrement but
each happening in a different class was making things less obvious. Also, the
fact that we were passing and grabbing references to mutable AtomicLong made it
harder that it ought to be to check where they were updated exactly.

For that to work properly, we have to create that OutChannel as soon as
handshake succeed and pass it to the MessageOutHandler so it can delegate to
that class when to flush. Which in turns change slightly how things needs to
be passed from OutboundMessagingConnection into
OutboundConnection/OutboundHandshakeHandler.

The patch also adds a OutboundConnectionIdentifier class that simply group the
local/remote/preferredRemote addresses as those were passed around a bit and
it felt a tad cleaner. It does makes it a bit more obvious when we actually use
the remove "identifier" address versus the "preferred" address though.

Another (important) thing that should be mentioned is that the code was previously
retrying timeout-ed message (on ExpiredException). I'm pretty sure this was incorrect
however: timeout-ed messages are ones for which no one is waiting for the
response and that we want to drop, so we shouldn't retry them. In fact, the code
was definitively not right on that front since while the patch was requeuing
expired message, it was doing so without updating the message timestamp or
timeout, so the message would be rejected as expired in MessageOutHandler again
leading to, unless I'm missing something, an infinite loop. Anyway, the patch
fixes that by simply dropping expired messages.

Lastly, I'll note that the patch slightly change the behavior of
cancelling a message promise future. I don't think we ever do that so it
probably doesn't matter much, but I would expect that if someone ever
does, what it means is "cancelling" the message itself, so not sending
the message if it hasn't been sent yet and that's all. And that's what
the patch does. Previously, cancelling a single message future was
trashing the connection and still (potentially) retrying the message.
This didn't felt right to me, so hopefully I haven't missed a good
reason why it was that way.