Remove DisabledCoalescingStrategy, using null instead when not coalescing
This might be a personal preference thing, but I found it somewhat
"dishonest" that we had a DisabledCoalescingStrategy but were actually
special casing all code to _not_ use the strategy when there was no
coalescing. So while I would than in many context avoiding null can be
preferable, I feel that it's actually cleaner/shows the intent more
clearly in this particular case. Basically, it saves a bit of code and
I don't think makes things really more confusing/dangerous.