Make SAI sstable components immutable
When a SAI index is rebuild, create new index files rather than
overwriting existing files, making SAI index components files
essentially truly immutable (once initially written).

To do this, SAI index components have a new parameter, their
"generation": for a given version, the component with the highest
generation is the most recently written one. That generation is added to
the filename, but with the special case that the initial generation (for
a version), generation 0, is special cased and not added to the
filename, and this for backward compatibility. This means in particular
that filenames are completely unchanged by this patch up until one
triggers an index rebuild that does not also correspond to a version
change.

One implication is that on a rebuild, old SAI index components are not
deleted anymore (previously, they were deleted at the end of the
rebuild, and the rebuild was re-creating new versions of the same files,
essentially overwriting the file but in 2 steps) and will remain on
disk. The intent here is that this pave the way for allowing index
"live" rebuild, that is rebuilding an index without making the index
un-queriable during the rebuild. However, this patch does *not*
implement such "live" rebuild, leaving it as a follow up. As of this
patch, the old index component files are also never deleted, though over
time, compaction will get rid of them by virtue of replacing the
sstables that have those components.

Technically, the generation added is not incremented "per component",
but rather per "group" of components, where "group" means (for a
particular sstable) either 1) all the per-sstable components or 2) all
the components of a particular index. As an example, suppose the
components for an sstable are (where 1 is the generation, and including
only a subset of real components to keep the example shorter):
```
<sstable X>-SAI+CA+1+GroupMeta.db
<sstable X>-SAI+CA+1+TokenValues.db
<sstable X>-SAI+CA+1+PrimaryKeyTrie.db
<sstable X>-SAI+CA+1+GroupComplete.db
<sstable X>-SAI+CA+myIndex+Meta.db
<sstable X>-SAI+CA+myIndex+TermsData.db
```
In this example, the per-sstable group is complete, but not the group
for `myIndex` (missing at least the `-ColumnComplete` component). Also,
the per-sstable group is at generation 1, while the `myIndex` group is
at generation 0.. If a rebuild is triggered, the created components
will be (assuming `CA` is the current version and again using a subset
of the actual components for simplicity):
```
<sstable X>-SAI+CA+2+GroupMeta.db
<sstable X>-SAI+CA+2+TokenValues.db
<sstable X>-SAI+CA+2+PrimaryKeyTrie.db
<sstable X>-SAI+CA+2+GroupComplete.db
<sstable X>-SAI+CA+1+myIndex+Meta.db
<sstable X>-SAI+CA+1+myIndex+TermsData.db
<sstable X>-SAI+CA+1+PostingLists.db
<sstable X>-SAI+CA+1+ColumnComplete.db
```
In other words, the generation for the per-sstable group and the
`myIndex` one are independent, but the generation are linked within a
group (the `<sstable X>-SAI+CA+1+PostingLists.db` file is at generation
1 like the rest of that group and not 0, despite there being no
generation 0 file before the rebuild).

The reason generation are linked "by group" is that 1) we need to at
least link the generation to the "completion marker" components to the
rest of the component of the group this is a marker for, and linking the
whole group as above feels like the cleaner solution and 2) linking the
generation of *all* the SAI components of a given sstable would prevent
us from being able to rebuild a single index without rebuilding the
per-sstable components/other indexes.